Clean up

+++ src/main/java/com/gameserver/model/actor/BaseActor.java
@@ -197,7 +197,7 @@ public abstract class BaseActor {
                     ((PlayableCharacter) this).sendPacket(new SystemMessage("You received "+((NPCActor) target).getBaseExperience()+" experience"));
                     ((PlayableCharacter) this).sendPacket(new SystemMessage("You current EXP:  "+((PlayableCharacter) this).getCurrentExperience()));
 
-                    new Task(new SpawnActorTask(((NPCActor) target)), (int) ((((NPCActor) target).getRespawnTime()) * 1000));
+                    new Task(new SpawnActorTask(((NPCActor) target)), (((NPCActor) target).getRespawnTime()) * 1000);
                     target.setDead(true);
                     target.getActorIntention().setIntention(new IntentionIdle());
                     World.getInstance().removeActorFromWorld(target);
